Start implementing fucionality
For now, we have VCF streams and filters that can read the data in a VCF
file. The VCF information is reads sequencially, with metadata and
headers being read at once and lines in the file read one at a time. The
VCFStream class handes this behaviour. VCFFilters can be used to read
from those streams of VCF data and manipulate them. Two basic filters
are included:

- Individuals filter, which keeps from the VCF information about only a
  subset of all the individuals in the file
- Polymorphism fitler, which only emits the data on certain
  polymorphisms, identifier by ID