Create predictors and scripts to run them
This commit creates a set of classes that handle prediction of an
individual's population, based on how its genotype compares with the
average genotypes of the individuals in all the populations (not direct
comparison of individuals). We've added two predictors:

- PredictorJoao predicts the population that is closest in allele
frequency to the one of the individual
- PredictorMariana compares all possible populations in pairs and each
pair produces a winner, with the most frequent winner being the final
prediction

A predictor is anythig that, given a genotype, produces a distribution
of probabilities for the possible labels: that is, it produces a mapping
from the possible labels to floating points between 0 and 1, where the
various values sum up to 1.